refactor: divide test_dsl_pipeline.py
